continue taking the lazy way out regarding rng syncing
I could have introduced a Mutex that blocked on multithreaded access, either in the rust code, or less usefully in the tests only. But I plan to only do stuff with the map generator on a single thread, so it doesn't seem to be worth the effort/extra dependencies. Maybe when std::lazy::OnceCell is stable?